Move the Request object to Processor namespace and renamed.
